refactor: replace legacy ioc_mcp with new implementation (#12)
- Replace basic ioc_mcp.py with enhanced version (ioc_mcp_new.py)
- Update service name 'ioc_new' -> 'ioc' and port 8889 -> 8888
- New version adds screenshot capture and markdown reports

Relates to #10, #11
